* compute_base::get_device: added support to select ANY device (which just defaults to the first valid device right now) and the FASTEST overall device (this is computed by each implementation and can return a gpu or cpu device)  * opencl_compute: compute the fastest device (using the same cpu/gpu score) and also fix a potential problem where fastest_*_device pointers might store invalid devices  * compute_base::init: platform_index is now an uint32_t and defaults to ~0u (all bits set) which means that it should look up that value in the floor config file (if it doesn't exist there, it defaults to 0); only relevant for opencl  * llvm_compute: minor cleanup, remove the init function and use of cudacl, which is no longer used/available and will be replaced by something better in the future (default to sm_20 for now)  * minor xcode project update (still had TCC_LIB_ONLY defines in there)  * now that the basic compute stuff is wokring (at least with opencl), bump version to v0.2.0-d2

+++ compute/cuda/cuda_compute.cpp
@@ -31,7 +31,7 @@
 #endif
 
 void cuda_compute::init(const bool use_platform_devices floor_unused,
-						const size_t platform_index floor_unused,
+						const uint32_t platform_index floor_unused,
 						const bool gl_sharing floor_unused,
 						const unordered_set<string> device_restriction floor_unused) {
 	// TODO: !
